+++ docs/source/conf.py
@@ -190,4 +190,5 @@ intersphinx_mapping = {
     'http://static.repoze.org/whodocs/': None,
     'http://code.gustavonarea.net/repoze.what.plugins.sql/': None,
     'http://code.gustavonarea.net/repoze.what.plugins.xml/': None,
+    'http://code.gustavonarea.net/repoze.what-pylons/': None,
     }